bravado.swagger_type is no more
