+++ bravado/mapping/unmarshal.py
@@ -3,8 +3,11 @@ import jsonschema
 from bravado.exception import SwaggerError
 from bravado.mapping import formatter, schema
 from bravado.mapping.model import is_model, MODEL_MARKER
-from bravado.swagger_type import SWAGGER20_PRIMITIVES, is_list_like, \
-    is_dict_like
+from bravado.mapping.schema import (
+    is_dict_like,
+    is_list_like,
+    SWAGGER20_PRIMITIVES
+)
 
 
 def unmarshal_schema_object(swagger_spec, schema_object_spec, value):
