+++ tests/mapping/docstring/formatted_type_test.py
@@ -1,14 +1,14 @@
 import pytest
 
 from bravado.exception import SwaggerError
-from bravado.swagger_type import get_swagger_type
+from bravado.mapping.docstring import formatted_type
 
 
 def test_type_only():
     schema_obj = {
         'type': 'integer',
     }
-    assert 'integer' == get_swagger_type(schema_obj)
+    assert 'integer' == formatted_type(schema_obj)
 
 
 def test_format_and_type():
@@ -16,7 +16,7 @@ def test_format_and_type():
         'type': 'integer',
         'format': 'int64',
     }
-    assert 'integer:int64' == get_swagger_type(schema_obj)
+    assert 'integer:int64' == formatted_type(schema_obj)
 
 
 def test_array():
@@ -26,17 +26,17 @@ def test_array():
             'type': 'string',
         }
     }
-    assert 'array:string' == get_swagger_type(schema_obj)
+    assert 'array:string' == formatted_type(schema_obj)
 
 
 def test_ref():
     schema_obj = {
         '$ref': '#/definitions/Foo',
     }
-    assert '#/definitions/Foo' == get_swagger_type(schema_obj)
+    assert '#/definitions/Foo' == formatted_type(schema_obj)
 
 
 def test_missing_type_raises_error():
     with pytest.raises(SwaggerError) as excinfo:
-        get_swagger_type({'TyP3': 'string'})
+        formatted_type({'TyP3': 'string'})
     assert 'No proper type' in str(excinfo.value)
